Better Enemy

+++ Assets/Scripts/Models/Blueprints/CastleConfigBlueprint.cs
@@ -2,7 +2,8 @@
 {
     using BlueprintFlow.BlueprintReader;
 
-    [BlueprintReader("CastleConfig", true)] [CsvHeaderKey("CastleConfigId")]
+    [BlueprintReader("CastleConfig", true)]
+    [CsvHeaderKey("CastleConfigId")]
     public class CastleConfigBlueprint : GenericBlueprintReaderByRow<string, CastleConfigRecord>
     {
     }
@@ -10,9 +11,9 @@
     public class CastleConfigRecord
     {
         public string CastleConfigId        { get; set; }
-        public int    BaseMP                { get; set; }
-        public int    BaseHP                { get; set; }
-        public int    BaseGoldNeedToUpgrade { get; set; }
+        public float  BaseMP                { get; set; }
+        public float  BaseHP                { get; set; }
+        public float  BaseGoldNeedToUpgrade { get; set; }
         public float  CoefficientMP         { get; set; }
         public float  CoefficientHP         { get; set; }
         public float  CoefficientGold       { get; set; }
